requirements.txt
yfinance
pandas
numpy
scikit-learn
xgboost
sqlalchemy
pyodbc
mysql-connector-python
